requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
anyio==4.13.0
asteroid-filterbanks==0.4.0
attrs==26.1.0
audioop-lts==0.2.2
audioread==3.1.0
av==17.0.1
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
ctranslate2==4.7.1
cycler==0.12.1
decorator==5.2.1
einops==0.8.2
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.29.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.47.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
onnxruntime==1.25.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-exporter-otlp-proto-http==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
optuna==4.8.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.6
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.49
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.11.0
torchcodec==0.11.1
torchmetrics==1.9.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
gunicorn
